+++ src/Tooling/DoctrineMigrations/Version20170805082413.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace App\Migrations;
+namespace Tooling\DoctrineMigrations;
 
-use Doctrine\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
 
-class Version20170805082413 extends AbstractMigration
+final class Version20170805082413 extends AbstractMigration
 {
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
         // We've updated the Prooph components, this method doesn't exist anymore.
         // Version 20180907111518 will create the new stream(s).
@@ -21,10 +21,11 @@ class Version20170805082413 extends AbstractMigration
         //     EventStoreSchema::createSingleStream($schema, 'event_stream', true);
         // }
     }
+
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         // if (class_exists('Prooph\EventStore\Adapter\Doctrine\Schema\EventStoreSchema')) {
         //     EventStoreSchema::dropStream($schema);
