+++ src/Tooling/DoctrineMigrations/Version20170811124421.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace App\Migrations;
+namespace Tooling\DoctrineMigrations;
 
+use Doctrine\DBAL\Schema\Schema;
 use Doctrine\DBAL\Schema\SchemaException;
 use Doctrine\Migrations\AbstractMigration;
-use Doctrine\DBAL\Schema\Schema;
 
-class Version20170811124421 extends AbstractMigration
+final class Version20170811124421 extends AbstractMigration
 {
     /**
      * @param Schema $schema
      *
      * @throws SchemaException
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
         $userTable = $schema->getTable('user');
         $userTable->addColumn('shepherd_id', 'shepherd_id');
@@ -27,7 +27,7 @@ class Version20170811124421 extends AbstractMigration
      *
      * @throws SchemaException
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         $userTable = $schema->getTable('user');
         $userTable->dropIndex('UNIQ_8D93D6493AE5C753');
