+++ src/Tooling/DoctrineMigrations/Version20171103163625.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace App\Migrations;
+namespace Tooling\DoctrineMigrations;
 
-use Doctrine\Migrations\AbstractMigration;
 use Doctrine\DBAL\Schema\Schema;
-use Elewant\Herding\Projections\HerdReadModel;
+use Doctrine\Migrations\AbstractMigration;
+use Elewant\Webapp\Infrastructure\ProophProjections\HerdReadModel;
 
-class Version20171103163625 extends AbstractMigration
+final class Version20171103163625 extends AbstractMigration
 {
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
         $desiredBreeds = $schema->createTable(HerdReadModel::TABLE_DESIRED_BREEDS);
         $desiredBreeds->addColumn('herd_id', 'guid');
@@ -27,7 +27,7 @@ class Version20171103163625 extends AbstractMigration
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         $schema->dropTable(HerdReadModel::TABLE_DESIRED_BREEDS);
     }
