+++ src/Tooling/DoctrineMigrations/Version20180907111518.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Migrations;
+namespace Tooling\DoctrineMigrations;
 
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\Migrations\AbstractMigration;
@@ -12,7 +12,7 @@ final class Version20180907111518 extends AbstractMigration
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
         $this->addSql(
             file_get_contents(__DIR__ . '/../../config/scripts/mariadb/01_event_streams_table.sql')
@@ -25,7 +25,7 @@ final class Version20180907111518 extends AbstractMigration
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         $schema->dropTable('event_streams');
         $schema->dropTable('projections');
