+++ src/Tooling/SymfonyConsole/CreateEventStreamCommand.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Command;
+namespace Tooling\SymfonyConsole;
 
+use ArrayIterator;
 use Prooph\EventStore\EventStore;
 use Prooph\EventStore\Stream;
 use Prooph\EventStore\StreamName;
@@ -13,21 +14,21 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 final class CreateEventStreamCommand extends ContainerAwareCommand
 {
-    protected function configure()
+    protected function configure(): void
     {
         $this->setName('event-store:event-stream:create')
             ->setDescription('Create event_stream.')
             ->setHelp('This command creates the event_stream');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): void
     {
         /** @var EventStore $eventStore */
         $eventStore = $this->getContainer()->get('prooph_event_store.herd_store');
 
         $streamName = new StreamName('event_stream');
         if (!$eventStore->hasStream($streamName)) {
-            $eventStore->create(new Stream($streamName, new \ArrayIterator([])));
+            $eventStore->create(new Stream($streamName, new ArrayIterator([])));
             $output->writeln('<info>Event stream was created successfully.</info>');
         }
     }
